% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_sdm.r
\name{benchmark_sdm}
\alias{benchmark_sdm}
\title{Benchmark regular machine learning models}
\usage{
benchmark_sdm(benchmarking_data, learners, dataset_type = "default",
  sample = FALSE)
}
\arguments{
\item{benchmarking_data}{Dataframe from the output of \code{\link{get_benchmarking_data}} function.}

\item{learners}{MLR learner objects}

\item{dataset_type}{Partitioning method}

\item{sample}{Logical value to indicate whether benchmarking should be done on an undersampled dataset}
}
\value{
Benchmarking object (bmr class)
}
\description{
Benchmark regular machine learning models
}
