% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_predict_functions.r
\name{customPredictFun}
\alias{customPredictFun}
\title{Custom functions for dismo prediction}
\usage{
customPredictFun(model, data)
}
\arguments{
\item{model}{MLR trained model.}

\item{data}{A dataframe containing occurence data.}
}
\value{
A vector with predictions.
}
\description{
A collection of functions that enable the usage of mlr model predictions with dismo.
}
