% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_dl.r
\name{evaluate_dl}
\alias{evaluate_dl}
\title{Evaluate deep learning model performance}
\usage{
evaluate_dl(model_keras, input_data)
}
\arguments{
\item{model_keras}{Keras deep learning model}

\item{input_data}{Data parsed in a suitable way by using the  \code{\link{prepare_dl_data}} function.}
}
\value{
A table containing the model estimates
}
\description{
Evaluate deep learning model performance
}
