% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_maxent.r
\name{evaluate_maxent}
\alias{evaluate_maxent}
\title{Evaluate MaxEnt model performance}
\usage{
evaluate_maxent(raster_data, method)
}
\arguments{
\item{raster_data}{A raster dataset}

\item{method}{Partitioning method.}
}
\value{
List containing AUC value and predict object (for plotting)
}
\description{
Evaluate MaxEnt model performance
}
