% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmarking_data.r
\name{get_benchmarking_data}
\alias{get_benchmarking_data}
\title{Download data for benchmarking}
\usage{
get_benchmarking_data(scientific_name, limit = 1000,
  bioclim_resolution = 10)
}
\arguments{
\item{scientific_name}{A character string indicating scientific species name.}

\item{limit}{A numeric value indicating maximum number of occurence records requested.}

\item{bioclim_resolution}{A numeric vlaue indicating the resolution of the raster Bioclim variables.}
}
\value{
A list containing the downloaded datasets.
}
\description{
Download data for benchmarking
}
\examples{
result.data <- get_benchmarking_data("Lynx lynx", limit = 1500, bioclim_resolution = 10)
}
