% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_model_results.r
\name{get_best_model_results}
\alias{get_best_model_results}
\title{Extract best model performances}
\usage{
get_best_model_results(bmr)
}
\arguments{
\item{bmr}{A bmr object, result of the \code{\link{benchmark_sdm}} function.}
}
\value{
A dataframe containing the best model results (learner name, optimal iteration number and associated AUC).
}
\description{
Extract best model performances
}
\examples{
best_results <- get_best_model_results(bmr)
}
