% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl_auc.r
\name{get_dl_auc}
\alias{get_dl_auc}
\title{Extract AUC from keras evaluation}
\usage{
get_dl_auc(keras_evaluation)
}
\arguments{
\item{keras_evaluation}{A table containing the keras evaluation.}
}
\value{
A numeric value indicating the AUC of the tested keras model.
}
\description{
Extract AUC from keras evaluation
}
\examples{
benchmarking_data <- get_benchmarking_data("Lynx lynx", limit = 1500, climate_resolution = 10)
benchmarking_data_dl <- prepare_dl_data(benchmarking_data$df_data, "default")
keras_results <- train_dl(benchmarking_data_dl)
keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)
keras_auc <- get_dl_auc(keras_evaluation)
}
