% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.r
\name{partition_data}
\alias{partition_data}
\title{Partition data in order to avoid spatial auto correlation}
\usage{
partition_data(dataset_raster, dataset, env, method)
}
\arguments{
\item{dataset_raster}{Raster data}

\item{dataset}{Dataframe containing species occurences}

\item{env}{Raster bioclim variables}

\item{method}{Desired partitioning method}
}
\value{
Dataframe partitionined using the selected method
}
\description{
Partition data in order to avoid spatial auto correlation
}
