% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdm_map.r
\name{plot_sdm_map}
\alias{plot_sdm_map}
\title{Plot SDM map}
\usage{
plot_sdm_map(raster_data, bmr_models, model_id, model_iteration)
}
\arguments{
\item{raster_data}{Raster data}

\item{bmr_models}{Models extracted from the benchmarking bmr object}

\item{model_id}{The model id of interest}

\item{model_iteration}{The model iteration of interesst}
}
\value{
SDM Map plot
}
\description{
Plot SDM map
}
