% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dl_data.r
\name{prepare_dl_data}
\alias{prepare_dl_data}
\title{Parse data for deep learning model training}
\usage{
prepare_dl_data(input_data, partitioning_type)
}
\arguments{
\item{input_data}{A dataframe containing the input data.}

\item{partitioning_type}{A character string indicating the desired spatial data partitioning method. Can be "default", "block", "checkerboard1", or "checkerboard2".}
}
\value{
A dataframe containing the prepared data.
}
\description{
Parse data for deep learning model training
}
\examples{
benchmarking_data <- get_benchmarking_data("Lynx lynx", limit = 1500, climate_resolution = 10)
benchmarking_data_dl <- prepare_dl_data(benchmarking_data$df_data, "default")
}
