% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sdmbench.r
\name{run_sdmbench}
\alias{run_sdmbench}
\title{Start sdmbench GUI in a browser window.}
\usage{
run_sdmbench()
}
\value{
A shiny app object
}
\description{
This interface can be used to use the complete package functionality without
writing code.
}
\examples{
\dontrun{
run_sdmbench()
}
}
