% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_predict_functions.r
\name{customPredictFun}
\alias{customPredictFun}
\title{Custom function for dismo prediction}
\usage{
customPredictFun(model, data)
}
\arguments{
\item{model}{MLR trained model}

\item{data}{Occurence data}
}
\value{
Vector with predictions
}
\description{
Custom function for dismo prediction
}
