% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_dl.r
\name{evaluate_dl}
\alias{evaluate_dl}
\title{Evaluate deep learning model performance}
\usage{
evaluate_dl(model_keras, input_data)
}
\arguments{
\item{model_keras}{Keras deep learning model.}

\item{input_data}{A dataframe containing occurence data parsed for deep learning using the \code{\link{prepare_dl_data}} function.}
}
\value{
A tibble containing the model estimates.
}
\description{
A function to evaluate a deep learning model.
}
\examples{
keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)
}
