% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_dl.r
\name{evaluate_dl}
\alias{evaluate_dl}
\title{Evaluate deep learning model performance}
\usage{
evaluate_dl(model_keras, input_data)
}
\arguments{
\item{model_keras}{Keras deep learning model.}

\item{input_data}{A dataframe containing occurrence data parsed for deep learning using the \code{\link{prepare_dl_data}} function.}
}
\value{
A tibble containing the model estimates.
}
\description{
A function to evaluate a deep learning model.
}
\examples{
# download benchmarking data
benchmarking_data <- get_benchmarking_data("Lynx lynx",
                                           limit = 1500,
                                           climate_resolution = 10)

# transform benchmarking data into a format suitable for deep learning
# if you have previously used a partitioning method you should specify it here
benchmarking_data_dl <- prepare_dl_data(input_data = benchmarking_data$df_data,
                                       partitioning_type = "default")

# perform sanity check on the transformed dataset
# for the training set
head(benchmarking_data_dl$train_tbl)
table(benchmarking_data_dl$y_train_vec)

# for the test set
head(benchmarking_data_dl$test_tbl)
table(benchmarking_data_dl$y_test_vec)

# train neural network
keras_results <- train_dl(benchmarking_data_dl)

# inspect results
keras_results$history

# you can also plot the results
plot(keras_results$history)

# create evaluation tibble containing deep learning training results
keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)
head(keras_evaluation)
}
