% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_maxent.r
\name{evaluate_maxent}
\alias{evaluate_maxent}
\title{Evaluate MaxEnt model performance}
\usage{
evaluate_maxent(raster_data, method)
}
\arguments{
\item{raster_data}{A raster dataset.}

\item{method}{A character string indicating the spatial data partitioning method.}
}
\value{
A list containing AUC value and predict object (for plotting).
}
\description{
Evaluate MaxEnt model performance
}
\examples{
benchmarking_data <- get_benchmarking_data("Lynx lynx", limit = 1500, climate_resolution = 10)
maxent_results <- evaluate_maxent(benchmarking_data$raster_data, method = "block")
}
