% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_model_results.r
\name{get_best_model_results}
\alias{get_best_model_results}
\title{Extract best model performances}
\usage{
get_best_model_results(bmr)
}
\arguments{
\item{bmr}{BMR object, result of the \code{\link{benchmark_sdm}} function.}
}
\value{
Dataframe containing the best model results
}
\description{
Extract best model performances
}
