% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl_auc.r
\name{get_dl_auc}
\alias{get_dl_auc}
\title{Extract AUC from keras evaluation}
\usage{
get_dl_auc(keras_evaluation)
}
\arguments{
\item{keras_evaluation}{A table containing the keras evaluation.}
}
\value{
A numeric value indicating the AUC of the tested keras model.
}
\description{
Extract AUC from keras evaluation
}
\examples{
\dontrun{
# download benchmarking data
benchmarking_data <- get_benchmarking_data("Lynx lynx",
                                           limit = 1500,
                                           climate_resolution = 10)

# transform benchmarking data into a format suitable for deep learning
# if you have previously used a partitioning method you should specify it here
benchmarking_data_dl <- prepare_dl_data(input_data = benchmarking_data$df_data,
                                       partitioning_type = "default")

# perform sanity check on the transformed dataset
# for the training set
head(benchmarking_data_dl$train_tbl)
table(benchmarking_data_dl$y_train_vec)

# for the test set
head(benchmarking_data_dl$test_tbl)
table(benchmarking_data_dl$y_test_vec)

# train neural network
keras_results <- train_dl(benchmarking_data_dl)

# inspect results
keras_results$history

# you can also plot the results
plot(keras_results$history)

# create evaluation tibble containing deep learning training results
keras_evaluation <- evaluate_dl(keras_results$model, benchmarking_data_dl)

# compute neural network AUC
keras_auc <- get_dl_auc(keras_evaluation)

# print AUC
keras_auc
}
}
