% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl_auc.r
\name{get_dl_auc}
\alias{get_dl_auc}
\title{Extract AUC from keras evaluation}
\usage{
get_dl_auc(keras_evaluation)
}
\arguments{
\item{keras_evaluation}{A table containing the keras evaluation.}
}
\value{
A numeric value indicating the AUC of the tested keras model.
}
\description{
Extract AUC from keras evaluation
}
\examples{
keras_auc <- get_dl_auc(keras_evaluation)
}
