% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.r
\name{partition_data}
\alias{partition_data}
\title{Partition spatial data}
\usage{
partition_data(dataset_raster, dataset, env, method)
}
\arguments{
\item{dataset_raster}{A raster dataset.}

\item{dataset}{A dataframe containing species occurrences.}

\item{env}{A raster dataset containing the environmental variables.}

\item{method}{A character string representing the desired spatial partitioning method. Can be "default", "block", "checkerboard1", or "checkerboard2".}
}
\value{
A dataframe partitioned using the selected method.
}
\description{
A function that partitions spatial data in order to avoid spatial autocorrelation.
}
\examples{
benchmarking_data <- get_benchmarking_data("Lynx lynx", limit = 1500, climate_resolution = 10)
benchmarking_data$df_data <- partition_data(dataset_raster = benchmarking_data$raster_data, dataset = benchmarking_data$df_data, env = benchmarking_data$raster_data$climate_variables, method = "block")
}
