% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dl_map.r
\name{plot_dl_map}
\alias{plot_dl_map}
\title{Plot SDM map from deep learning predictions}
\usage{
plot_dl_map(raster_data, keras_model, custom_fun)
}
\arguments{
\item{raster_data}{Raster dataset containing the data}

\item{keras_model}{Trained deep learning model}

\item{custom_fun}{Custom predict function to enable the prediction}
}
\value{
Plot of SDM
}
\description{
Plot SDM map from deep learning predictions
}
