% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdm_map.r
\name{plot_sdm_map}
\alias{plot_sdm_map}
\title{Plot SDM map}
\usage{
plot_sdm_map(raster_data, bmr_models, model_id, model_iteration,
  map_type = "static")
}
\arguments{
\item{raster_data}{A raster dataset containing the occurrence data.}

\item{bmr_models}{A list of models extracted from the benchmarking bmr object.}

\item{model_id}{A character string indicating the model id of interest.}

\item{model_iteration}{A numeric value indicating the model iteration of interest.}

\item{map_type}{A logical indicating if the map should be static or interactive.}
}
\value{
An interactive leaflet map, showing the species distribution.
}
\description{
Plot SDM map
}
\examples{
# download benchmarking data
benchmarking_data <- get_benchmarking_data("Lynx lynx",
                                           limit = 1500,
                                           climate_resolution = 10)

# create a list of learners (algorithms) to compare
learners <- list(mlr::makeLearner("classif.randomForest",
                                  predict.type = "prob"),
                 mlr::makeLearner("classif.logreg",
                                  predict.type = "prob"))

# run the benchmarking
bmr <- benchmark_sdm(benchmarking_data$df_data,
                     learners = learners,
                     dataset_type = "default",
                     sample = FALSE)

# get best model results
# you should obtain a dataframe containing the highest performing (by AUC)
# algorithm name, iteration and associated AUC
best_results <- get_best_model_results(bmr)

# plot the SDM map of the best performing model
plot_sdm_map(raster_data = benchmarking_data$raster_data$climate_variables,
             bmr_models = bmr$learners,
             model_id = best_results$learner.id[1],
             model_iteration = best_results$iter[1],
             map_type = "static")
}
