% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdm_map.r
\name{plot_sdm_map}
\alias{plot_sdm_map}
\title{Plot SDM map}
\usage{
plot_sdm_map(raster_data, bmr_models, model_id, model_iteration)
}
\arguments{
\item{raster_data}{A raster dataset containing the occurence data.}

\item{bmr_models}{A list of models extracted from the benchmarking bmr object.}

\item{model_id}{A character string indicating the model id of interest.}

\item{model_iteration}{A numeric vlaue indicating the model iteration of interest.}
}
\value{
A plot, showing the species distribution.
}
\description{
Plot SDM map
}
\examples{
\dontrun{
plot_sdm_map(raster_data = benchmarking_data$raster_data$climate_variables, bmr_models = bmr_models, model_id = best_results$learner.id[4], model_iteration = best_results$iter[4])
}
}
