% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dl_data.r
\name{prepare_dl_data}
\alias{prepare_dl_data}
\title{Parse data for it to be suitable for deep learning modeling}
\usage{
prepare_dl_data(input_data, partitioning_type)
}
\arguments{
\item{input_data}{Input dataframe}

\item{partitioning_type}{Desired data partitioning type}
}
\value{
Parsed dataframe
}
\description{
Parse data for it to be suitable for deep learning modeling
}
