% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_dl.r
\name{train_dl}
\alias{train_dl}
\title{Train deep learning model}
\usage{
train_dl(input_data)
}
\arguments{
\item{input_data}{Parsed deep learning dataframe}
}
\value{
List containing the trained deep learning model, and a history object for performance evaluation
}
\description{
Train deep learning model
}
